attempt to modularise
