+++ nbstripout/__init__.py
@@ -1,0 +1,4 @@
+from ._nbstripout import install, uninstall, status, main
+from ._utils import pop_recursive, strip_output
+__all__ = ["install", "uninstall", "status", "main",
+           "pop_recursive", "strip_output"]
